requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bleach==6.1.0
bootstrap4==0.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cssmin==0.2.0
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-htmlmin==0.11.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
html5lib==1.1
humanize==4.9.0
idna==3.6
importlib-metadata==6.7.0
mysqlclient==2.2.1
pillow==10.2.0
python-memcached==1.62
pytz==2024.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
zipp==3.15.0
